+++ src/ubuntu/install/zoom/install_zoom.sh
@@ -1,9 +1,18 @@
 #!/usr/bin/env bash
 set -ex
-wget -q  https://zoom.us/client/latest/zoom_amd64.deb
+
+ARCH=$(arch | sed 's/aarch64/arm64/g' | sed 's/x86_64/amd64/g')
+
+if [ "${ARCH}" == "arm64" ] ; then
+    echo "Zoom for arm64 currently not supported, skipping install"
+    exit 0
+fi
+
+
+wget -q  https://zoom.us/client/latest/zoom_${ARCH}.deb
 apt-get update
 apt-get install -y maximus
-apt-get install -y ./zoom_amd64.deb
+apt-get install -y ./zoom_${ARCH}.deb
 rm zoom_amd64.deb
 cp /usr/share/applications/Zoom.desktop $HOME/Desktop/
 chmod +x $HOME/Desktop/Zoom.desktop
